+++ packages/@rescript/runtime/lib/js/Stdlib_TaggedTemplate.cjs
@@ -1,0 +1,7 @@
+'use strict';
+
+
+let make = (transform => (strings, ...values) => transform(strings, values));
+
+exports.make = make;
+/* No side effect */